+++ src/main/java/chess/ConsoleApplication.java
@@ -1,5 +1,7 @@
 package chess;
 
+import chess.controller.ChessController;
+
 public class ConsoleApplication {
 
     public static void main(String[] args) {
