+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,0 +1,24 @@
+package chess.domain.piece;
+
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
+import chess.utils.PossiblePositionChecker;
+
+import java.util.List;
+
+public class Knight extends Piece {
+
+    public Knight(Color color) {
+        super(Type.KNIGHT, color);
+    }
+
+    @Override
+    public boolean isMovable(Position source, Position target) {
+        return PossiblePositionChecker.isMovablePositions(Direction.knight(), source, target);
+    }
+
+    @Override
+    public List<Position> computeBetweenTwoPositionByLine(Position source, Position target) {
+        return List.of();
+    }
+}
