+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,7 +1,7 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.position.Direction;
-import chess.position.Position;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
 
 import java.util.List;
