+++ src/main/java/chess/domain/position/Direction.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 import java.util.List;
