+++ src/main/java/chess/domain/state/State.java
@@ -1,8 +1,8 @@
-package chess.state;
+package chess.domain.state;
 
-import chess.Chessboard;
-import chess.piece.Color;
-import chess.position.Position;
+import chess.domain.board.Chessboard;
+import chess.domain.piece.Color;
+import chess.domain.position.Position;
 
 public interface State {
