+++ src/main/java/chess/utils/BetweenPositionsGenerator.java
@@ -1,6 +1,6 @@
 package chess.utils;
 
-import chess.position.Position;
+import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
