+++ src/test/java/chess/StateTest.java
@@ -1,8 +1,9 @@
 package chess;
 
-import chess.state.Finish;
-import chess.state.Play;
-import chess.state.Ready;
+import chess.domain.Turn;
+import chess.domain.state.Finish;
+import chess.domain.state.Play;
+import chess.domain.state.Ready;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
