+++ src/test/java/chess/piece/PawnTest.java
@@ -1,6 +1,8 @@
 package chess.piece;
 
-import chess.position.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Pawn;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
