+++ src/test/java/chess/piece/RookTest.java
@@ -1,6 +1,8 @@
 package chess.piece;
 
-import chess.position.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Rook;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
